requirements.txt
opencv-python
yt-dlp
ultralytics
numpy
scipy
pycryptodome
requests
